Create README.md
For my first front end development project, I have made a website for my husband who is a local standup comedian.  
I have made a navigation bar that includes 2 media queries. The full menu disappears and gets smaller while a hamburger menu appears when it goes from desktop to tablet to phone. Javascript is used twice. Once to open up the hamburger menu when it's mobile. There is also another JavaScript button that responds with the punchline to a joke.   
I have used CSS FlexBox to organize the menu bar and make it responsive. 
There are more than 5 commits.